+++ terra-rust-api/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.2
 ### 0.2.10 - devel
+* add Eq/ToString to NodeIpPort
 ### 0.2.9 - 18-Aug-21
 * Add slashing/Unjail
 * Add market/Swap